pyproject.toml
[project]
name = "fastapi-mvc-app"
version = "0.1.0"
description = "FastAPI MVC application"
authors = [
    { name = "Nazar Znetyniak", email = "72510445+NazarZnet@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "bcrypt (<4.1)",
    "redis (>=5.2.1,<6.0.0)",
    "fastapi-redis-cache (>=0.2.5,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
